Made all currencies into doubles to make smaller values possible

+++ src/main/java/com/petometry/currencyservice/rest/model/BalancesDto.java
@@ -7,5 +7,5 @@ import lombok.ToString;
 @ToString
 public class BalancesDto {
 
-    private Long geocoin;
+    private Double geocoin;
 }
